fixed imports

+++ src/program/src/main/scala/filters/MaxPriceFilter.scala
@@ -1,7 +1,7 @@
 package filters
 
-import comparator.{Filter, Offer}
+import comparator.Offer
 
-class MaxPriceFilter(val maxPrice: Int) extends Filter {
+class MaxPriceFilter(val maxPrice: Double) extends Filter {
     override var filter: Offer => Boolean = o => o.price <= maxPrice
 }
